@@ -6,11 +6,11 @@
     "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
-    "react": "15.0.2",
-    "react-native": "0.26.0",
+    "react": "^15.1.0",
+    "react-native": "^0.27.2",
     "react-native-db-models": "^0.1.3",
     "react-native-navbar": "^1.5.0",
     "react-native-side-menu": "^0.19.0",
-    "tcomb-form-native": "^0.5.0"
+    "tcomb-form-native": "^0.5.1"
   }
 }
