@@ -6,80 +6,63 @@
  */
 'use strict';
  
-/* ==============================
-  Initialise Component
-  =============================== */
-  // React
-  import React, { Component } from 'react';
-  import {
-    StyleSheet,
-    View,
-    Text,
-    TouchableOpacity,
-  } from 'react-native';
+/* Setup ==================================================================== */
+import React, { Component } from 'react'
+import {
+  StyleSheet,
+  View,
+  Text,
+  TouchableOpacity,
+} from 'react-native'
 
-  // App Globals
-  import AppStyles from '../styles.ios';
-  // import AppConfig from '../config.ios';
+// App Globals
+import AppStyles from '../styles.ios'
 
-  // Screens / Pages
-  // import AnotherPage from './tabbar.ios';
+// Components
+import Button from '../components/button.ios'
 
-  // Components
-  import Button from '../components/button.ios';
+/* Component ==================================================================== */
+class ComingSoon extends Component {
+  static propTypes = {
+    navigator: React.PropTypes.object.isRequired,
+  }
 
-/* ==============================
-  View
-  =============================== */
-  var ComingSoon = React.createClass({
+  /**
+    * Navigates to same scene (for Demo purposes)
+    */
+  _navigate = (navbarTitle) => {
+    this.props.navigator.push({
+      title: navbarTitle, 
+      component: ComingSoon, 
+      index: 2
+    });
+  }
 
-    /**
-      * Navigates to page from menu
-      */
-    _navigate: function(navbar_title) {
-      this.props.navigator.push({
-        title: navbar_title, 
-        component: ComingSoon, 
-        index: 2
-      });
-    },
+  /**
+    * RENDER
+    */
+  render = () => {
+    let text = this.props.placeholder || 'Coming soon...'
 
-    /**
-      * RENDER
-      */
-    render() {
-      var placeholderText = this.props.placeholder;
-      if(!placeholderText) { placeholderText = 'Coming soon...'; }
+    // Done
+    return (
+      <View style={[AppStyles.container, AppStyles.containerCentered]}>
+        <Text style={[AppStyles.baseText, AppStyles.p]}>
+          {text}
+        </Text>
 
-      // Done
-      return (
-        <View style={[AppStyles.container, AppStyles.containerCentered]}>
-          <Text style={[AppStyles.baseText, AppStyles.p]}>
-            {placeholderText}
-          </Text>
+        <View style={[AppStyles.spacer_10]} />
 
-          <View style={[AppStyles.spacer_10]} />
+        <Button type={'outlined'}
+          text={'Tap to test the back button'}
+          onPress={()=>this._navigate(text)} />
+      </View>
+    );
+  }
+}
 
-          <Button
-            text={'Tap to test the back button'}
-            style={'outlined'}
-            onPress={()=>this._navigate(placeholderText)} />
-        </View>
-      );
-    }
-
-  });
-
-/* ==============================
-  Styles
-  =============================== */
-  /*var styles = StyleSheet.create({
-  });*/
-
-/* ==============================
-  Done!
-  =============================== */
-  module.exports = ComingSoon;
-  module.exports.details = {
-    title: 'ComingSoon'
-  };
+/* Export Component ==================================================================== */
+module.exports = ComingSoon;
+module.exports.details = {
+  title: 'ComingSoon'
+};
